+++ packages/org/public/forms/org_employeeonstaff-fm.vue
@@ -24,17 +24,22 @@
         :label="'mi_dateFrom'"
         attribute-name="mi_dateFrom"
       />
-      <u-auto-field
-        :label="'mi_dateTo'"
+      <u-form-row
         attribute-name="mi_dateTo"
-      />
+        :label="'mi_dateTo'"
+      >
+        <u-date-picker
+          v-model="mi_dateTo"
+          type="datetime"
+        />
+      </u-form-row>
     </u-form-container>
   </div>
 </template>
 
 <script>
 const { Form, mapInstanceFields } = require('@unitybase/adminui-vue')
-const fieldList = ['ID', 'employeeID', 'staffUnitID', 'tabNo', 'employeeOnStaffType', 'description', 'mi_dateFrom',
+const fieldList = ['ID', 'employeeID', 'employeeID.fullFIO', 'staffUnitID', 'tabNo', 'employeeOnStaffType', 'description', 'mi_dateFrom',
   'mi_dateTo']
 const { required } = require('vuelidate/lib/validators/index')
 const { mapGetters } = require('vuex')
@@ -70,9 +75,25 @@ module.exports.default = {
   name: 'org_employeeonstaff',
   inject: ['entity', '$v'],
 
+  mixins: [
+    require('@unitybase/dom/public/mixins/form-caption-mixin')
+  ],
+
   computed: {
     ...mapInstanceFields(fieldList),
-    ...mapGetters(['loading'])
+    ...mapGetters(['loading']),
+
+    mi_dateTo: {
+      get () {
+        if (this.$store.state.data.mi_dateTo?.getFullYear() === 9999) {
+          return null
+        }
+        return this.$store.state.data.mi_dateTo
+      },
+      set (value) {
+        this.$store.commit('SET_DATA', { key: 'mi_dateTo', value })
+      }
+    }
   },
 
   methods: {